--- a/src/Combine.elm
+++ b/src/Combine.elm
@@ -39,7 +39,6 @@ module Combine exposing
     , sepBy1
     , sepBy1WithState
     , succeed
-    , succeedLazy
     , withState
     , withStateFromCore
     )
@@ -184,14 +183,6 @@ problem m =
 succeed : a -> Parser state a
 succeed res =
     Parser <| \state -> Core.succeed ( state, res )
-
-
-succeedLazy : (() -> a) -> Parser state a
-succeedLazy createRes =
-    Parser
-        (\state ->
-            Core.succeed ( state, createRes () )
-        )
 
 
 end : Parser state ()