… src/Elm/Parser/File.elm
… src/Elm/Parser/Imports.elm
… src/Elm/Parser/Layout.elm
… src/Elm/Parser/Modules.elm
… src/Elm/Parser/Node.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/State.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/Elm/Parser/Typings.elm
… src/Parser/Extra.elm
… tests/Elm/Parser/CommentTest.elm
… tests/Elm/Parser/DeclarationsTests.elm
… tests/Elm/Parser/FileTests.elm
… tests/Elm/Parser/LayoutTests.elm
… tests/Elm/Parser/Samples.elm
… tests/Elm/Parser/TokenTests.elm
… tests/Elm/Parser/TypeAnnotationTests.elm
… tests/Elm/Parser/TypingsTests.elm