WIP
- before: 42913.90000000596 ms to do 100 runs => 429.1390000000596 ms/run
- after: 31419.200000000186 ms to do 100 runs => 314.1920000000019 ms/run

done by
- ~more agressive pre-definitions and converting `shared |> Result.andThen (\partial -> succeed (with partial) ...)` to `map (\partial -> \with -> with partial) shared |> keep ...)`~ use intermediate union types for parsers that parse the rest of some structures
- convert `succeed` then drop/keep `parser` in pipelines to `map (\parsed -> ...) parser` (this was pretty fruiitful actually)
- introducing Combine.maybe(Map) as a more performant `oneOf [..., succeed ]`
- switch from `indents : List Int` in `State` to ~`indent : Int`~ `Core.getIndent` and `Core.withIndent`

I can think of some uglier ones if we want to get desperate.