@lue-bird 2024-07-24
Before I attempt this I'd like to check in with you:
I'd like to try eliminate
```elm
Combine.Parser a =
    ({ indents : List Int
     , comments : List (Node String)
     }
     ->
     Core.Parser
        ( { indents : List Int
          , comments : List (Node String)
          }
          , a
        )
    )
```
in favor of
```elm
ParserWithComments a =
    Core.Parser { comments : Rope (Node String), syntax : a }
```
using indentation handling built into `Core.Parser`

## upsides
- better performance (that's the hope anyway) due to less state and less functions
- no more mixing the two parser kinds, shared helpers between parser representations
- it's pretty easy to follow where information comes from ("no implicit state")
- if we in v8 decide to attach comments in the AST directly, this becomes trivial (more like it would _simplify_ the code)

## downsides
- less control over the indentation state, which means getting the parsed result + indentation in one go isn't possible
- combining comments in the right order is something parser devs have to explicitly get right everytime layout is used