@jfmengels 2024-07-25
This was amazing @lue-bird, thank you so much for your recent PRs. Performance has been going :rocket: with all these changes

![](https://cloud.githubusercontent.com/assets/170270/13374524/b674ebd8-ddb9-11e5-9c5e-c59a755b14a6.gif)

---

> Before I attempt this I'd like to check in with you, I'd like to try eliminate...

Am I understanding correctly that indentation is now solved through the more recent commits? Or would you like to reintroduce `indents` still?

> no more mixing the two parser kinds, less conversion, more shared helpers between parser representations

This would be good, because this optimization has been pretty rough for the project in terms of consistency of code patterns :sweat_smile: 

> So.. does this experiment sound reasonable or is there something to consider which could make this a bad change?

I have little insight into the repercussions of such a change. I'm all for trying it out though. **Go ahead!**

(In the meantime, I'll rebase the v8 branch from #230 on top of `master`)

> accumulating in a Rope adds a bit of overhead and makes e.g. sorted insert (see port documentation comment) harder

For the port documentation comment, that problem will go away with v8, so I don't mind it making the parser slower in the meantime (especially since ports are not that common).

> And should this rather be a separate PR?

Yes. I've merged it anyway, so no choice anymore :smile: 
But yes, explorations that may fail should usually be done in new PRs, so that we don't block work that we know is an improvement from getting merged.